Enable automatic test termination
This enables relocating the HTIF by setting the address in the configuration file.
Example:

    htif_base_addr: 0x80001000,

riscv-tests uses a range of addresses (see, say, nm
rv64ud-p-move|egrep " tohost").

We further had to change how memory segments are looked up as the
common value overlaps with simulated memory so it's important we
search that last.  Finally, the tests do not write the command
register so we trigger on the data register as well.