[RTLMAX-235] Add support for running the riscv-tests
Add makefile with rules to run the riscv-test under riscvemu
and spike, comparing the results of the 2 simulators is deferred
for later.

Add test_report.py script that parses the output status of the
riscvemu and generates a junit xml file that will be uploaded
to Jenkins or any other test-tracking status database.