[ARCHSIM-79] Add support for hanlding the linux boot-event
Add support for:
1) Handling the linux boot event passed in the validation CSR-1
2) Add command line option to terminate simulation when specific
events are encountered. (Currently only the linux boot)

This will allow us to write tests where we valiate under riscvemu
that a given kernel-boots and gracefully end the simulation (and not
rely on flaky string matching of output.

It will also be used for slicing the boot as well, and finding the
linux-boot instruction count