New Cosim API [RTLMAX-290] [RTLMAX-116]
Introducing the two-entrypoint and two-file Cosimulation API.  The
public interface is in riscvemu_cosim.h and the library is in
libriscvemu_cosim.a.

The usage is remarkably simple: set up state with
riscvemu_cosim_init() and advance it with riscvemu_cosim_step().  The
latter will be in charge of checking consistency with the model.