Pay down technical debt -- massive clean up
The original RISCVEMU supports a lot of options that we don't and
can't support.  Let's cut the cruft and fix the options that aren't
actually options.  This deleted a lot of code.

Update the README.md to actually reflect what this is.

Remove the riscvemu binaries and rename vharness to riscvemu.

Only support RV64GC