[ARCHSIM-90] handle_dut_overrides csrw should be ignored.
handle_dut_overrides is expected to coerce from the DUT
the results of csrr and update the destination gpr.
csrw has the same opcode as csrr but rd is zero. Calling
riscv_set_reg and trying to set x0 rightlfully generates an
assert. csrw should be silent and should have no effect.