bootrom code to match the DUT
This commit updates the bootrom code to match what is used by the DUT
when cosimulating. The DUT resets with the debug mode to be set, and
the updated bootrom causes the processor to jump to the test start by
manipulating the debug state (dpc,dcsr) as opposed to jumping indirectly.

NOTE: I think should not matter for riscvemu but longterm it would be great
if we could pass a binary file for the bootrom code as opposed to hardcoding
this in riscvemu as the bootrom code might change based on the platform we are
simulating.