Properly propagate termination rather than exit(0) [VERIF-176]
Obvious fact: libraries should never exit() out.

This change is a minimal fix to stop riscvemu from exit()ing upon
writes of validation events.  Instead it sets a new flag that is
supposed to be checked at the outer iterations.

Note, there's some confusion in the code re. the "power_down_flag".
It does not mean termination.  Added a comment to point this out.

In subsequent changes, we'll eliminate more exit()s.