[RTLMAX-286] Fixes to riscvemu to support checkpointing benchmarks
1.  Add argument to pass additional kernel boot arguments

Add option cmdline where we pass additional arguments to the linux
kernel.

2. Add new valiation events for benchmarks.

Add new validation events to mark the beginning and end
of a benchmark.
The linux events have been renamed for clarity.

These events are written to valiation CSR1 from userspace
apps/utils and the simulator can terminate and create
checkpoints if requested from the command line

Fix bugs:
-- Generating error for unknown event name passed in the command line

3.  Terminate execution properly

4. Terminate riscvemu when terminate_simulation flag is true

This patch removes a loop that was at the end of main in vharness
and was executing additional instructions after we have hit a reason
to terminate the simulation, either because we have hit a terminating
event (e.g. linux-boot-end or benchmark-start) or written to validation
csr0

Fix: m->trace is initialized to UINT64_MAX and the removed loop was becoming
practically an infinate one.

5. Fix checkpointing for non-machine code.

Fix bug in checkpoint creation that prevents creating checkpoints
while user mode.