[RTLMAX-353] Add new validation event to pass exit code.
Add a new event that allows us to write the exit code (or generic
uint32_t value) from the user-space and save it in the simulator.
When the simulation ends we check the value and if it is non-zero
then we generate an error-message and exit riscvemu with a non-zero
value.