Don't skip dcsr writes + emit raised exceptions in the trace
Addition of PMP invalidated the previous trace.  Chasing this and a
different bug, the fragility of depending on minstret became clear and
instead we dig into the hart of RISCVEMU and rely on
pending_exception.  Since we are here, we can dump this which is very
useful when reading traces.

It looks like this:

  ...
  0 0x00000000800001fc (0x0ff0000f)
  0 0x0000000080000200 (0x00100193) x 3 0x0000000000000001
  0 0x0000000080000204 (0x00000073) exception 8, tval 0000000000000000
  3 0x0000000080000004 (0x34202f73) x30 0x0000000000000008
  3 0x0000000080000008 (0x00800f93) x31 0x0000000000000008
  ...

which is compatible with the parser in riscvemu_cosim_test

Also, riscvemu_cosim_test never failed!