Fix termination on HTIF (tohost) writes
The mechanism RISCVEMU uses for HTIF handling depends on setting up a
seperate memory region for HTIF and handling it specially.  This is
problematic as the HTIF region overlaps with regular memory and in
fact this breaks down when using ELF binaries as regions are created
in a different order.