Implement debug mode [ARCHSIM-89] [RTLMAX-261]
Access to CSRs dcsr, dpc, and dscratch, as well as execution of dret
is now only legal in debug mode (which is turned off upon dret).

Furthermore, we expose a new API - riscv_set_debug_mode() control
this.

Note, RISCVEMU starts with in debug mode because out bootroom depends
on it.

Minor fix: by defaulting to a breakpoint in tdata1[0], the test
rv64mi-p-breakpoint now passes.