[hpm] Add mcountinhibit CSR from Priv 1.11 update.
This adds a 32-bit machine-level read/write CSR. The only special-feature is
mcountinhibit[1] is always 0 (time count is always enabled and cannot be
inhibited).

HOWEVER: this commit does NOT actually block the increment of instret or rdcycle
or any of the HPMs. This means that mcountinhibit tests will FAIL on standalone
RISCVEMU, but pass in cosimulation since rdtime and rdinstret are overridden
from the DUT.