Provide support for simultaneous interrupt and exceptions [ARCHSIM-301]
On Maxion, the interrupt can race the exception, that is, we can
observe architectural state from the exception while the 1st committed
instruction is from handling the interrupt (yes, this is a lovely
subtle issue).  We handle this by informing RISVEMU of all exceptions,
not just interrupts.  The only case that is handled different from
before is the case where we have both types pending in which case we
advance the simulation into the exception, hopefully reproducing it,
before raising the interrupt.

This is an extremely rare event.