[sting] Code is also subject to PMP restrictions
For performance, RISCVEMU uses a software TLB which means that we
generally bypass the PMP checks.  This change consults the PMP and
refuses to install a translation that could potentially fail the PMP
checks.  It also fixes a long standing hole in how PMP regions are
checked (turns out, checking for interval intersection is pretty
cheap).

This was exposed by
http://sc-jenkins01:8080/job/maxion/job/Sting/job/sting-regression/34/testReport/junit/(root)/sting-regression/maxion_verisim_inst_acc_fault_mexe_pmp_regn_locked_na4_nxxwxr_conf/
the inst_acc_fault_mexe_pmp_regn_locked_na4_nxxwxr_conf test (and a few more like it).