Let the ELF file set the memory base
It was awkward to have to manually set --memory_base_addr for every
elf file when it could be derived from the ELF.  Unfortunately, for
hystorical reasons, the plumming for this is a bit ugly.

Changed the handling of command line elf argument to create a fake
config structure for this and consolidate the handling of the "BIOS".
XXX Doesn't derive memory base address for the case where the ELF is
given in a real .cfg file - that would require more yet changes.